+++ rpbp/reference_preprocessing/label_orfs.py
@@ -1,8 +1,15 @@
 #! /usr/bin/env python3
 
+"""This script labels the ORFs based on their exon
+transcript structure with respect to annotated coding sequences
+
+Contains:
+    match_pairs
+    assign_transcripts
+"""
+
 import argparse
 import logging
-import tqdm
 
 import pandas as pd
 
@@ -15,36 +22,25 @@ logger = logging.getLogger(__name__)
 
 default_num_cpus = 1
 
-default_annotated_exons = None
-default_label_prefix = ""
-default_nonoverlapping_label = None
 
-
-def change_orf_id(row, match_pairs):
-    """Make sure the label matches the transcript part of the
-    orf id, i.e. we want the label to be consistent with the
-    transcript to which the ORF is associated."""
+def match_pairs(row, matched_pairs):
+    """Add transcript id to ORF id, so that the label is
+    consistent with the transcript to which the ORF is associated."""
 
     orf_id = row.id
-    if row.is_duplicated:
-        split_id = orf_id.rpartition('_')
-        transcript_id = split_id[0]
-        if (transcript_id, orf_id) not in match_pairs:
-            # just pick the first one, all associated transcripts
-            # share the same relationship with this ORF
-            transcript_id = [pair[0] for pair in match_pairs if pair[1] == orf_id][0]
-            orf_id = transcript_id + ''.join(split_id[1:])
+    # pick first match arbitrarily
+    transcript_id = [pair[0] for pair in matched_pairs if pair[1] == orf_id][0]
 
-    return orf_id
+    return str(transcript_id + '_' + orf_id)
 
 
-def check_orf_ids(orf_group, type_to_matches):
+def assign_transcripts(orf_group, type_to_matches):
 
     orf_type = orf_group['orf_type'].iloc[0]
-    match_pairs = type_to_matches[orf_type]
+    matched_pairs = type_to_matches[orf_type]
 
-    orf_group['id'] = orf_group.apply(change_orf_id,
-                                      match_pairs=match_pairs,
+    orf_group['id'] = orf_group.apply(match_pairs,
+                                      matched_pairs=matched_pairs,
                                       axis=1)
 
     return orf_group
@@ -52,37 +48,42 @@ def check_orf_ids(orf_group, type_to_matches):
 
 def main():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-                                     description='''Label the ORFs based on their exon structure 
-                                     wrt the annotated transcripts. The extracted_orfs input 
-                                     file can be overwritten.''')
+                                     description='''Label the ORFs based on their transcript
+        exon structure wrt the annotated transcripts.''')
+
     parser.add_argument('annotated_transcripts', help='''The annotated transcripts for the genome
         in BED12+ format.''')
+
     parser.add_argument('extracted_orfs', help='''The ORFs extracted from the transcripts 
         in BED12+ format.''')
-    parser.add_argument('orf_exons', help='''The exon blocks for the ORFs, in BED6+ format, obtained
-        from "split-bed12-blocks".''')
+
     parser.add_argument('out', help='''The output (BED12+.gz) file.''')
-    parser.add_argument('-p', '--num-cpus', help='''The number of CPUs to use to perform
-        BED operations.''', type=int, default=default_num_cpus)
-    parser.add_argument('-f', '--filter', help='''If this flag is given, then ORFs
-        which are completely covered by an annotated transcript are discarded. Use to filter 
-        uninteresting ORFs from a de novo assembly.''', action='store_true')
+
     parser.add_argument('-e', '--annotated-exons', help='''The annotated transcript 
         exons can be passed with this option. If they are not given, they will be 
-        split from the annotated transcripts.''', default=default_annotated_exons)
+        split from the annotated transcripts.''', default=None)
+
+    parser.add_argument('-o', '--orf-exons', help='''The exon blocks for the ORFs, in BED6+ format, 
+        obtained from "split-bed12-blocks". If they are not given, they will be split from the
+        extracted ORFs.''', default=None)
+
     parser.add_argument('-n', '--nonoverlapping-label', help='''If this option is given, 
         then the ORFs which do not overlap the annotated transcripts at all will be given this label.
         By default, remaining oof overlapping ORFs are assigned the "overlap" label.
         If not given, the ORFs outside of annotated regions are labeled as "suspect".''',
-                        default=default_nonoverlapping_label)
+                        default=None)
+
     parser.add_argument('-l', '--label-prefix', help='''This string is prepended to all labels 
         assigned to ORFs, e.g. to indicate ORFs from a de novo assembly (Rp-Bp assigns the label
         "novel" to these, however the string is not prepended to "canonical ORFs").''',
-                        default=default_label_prefix)
-    parser.add_argument('-s', '--skip-check', help='''If this flag is given, then the 
-        transcript part of the orf_id is not checked for consistency with the label. Use
-        when the orf_id is not associated with a transcript_id, such as in a de novo assembly.
-        If given, then the ORF exons file will be re-written accordingly.''', action='store_true')
+                        default='')
+
+    parser.add_argument('-f', '--filter', help='''If this flag is given, then ORFs
+        which are completely covered by an annotated transcript are discarded. Use to filter 
+        uninteresting ORFs from a de novo assembly.''', action='store_true')
+
+    parser.add_argument('-p', '--num-cpus', help='''The number of CPUs to use to perform
+            BED operations.''', type=int, default=default_num_cpus)
 
     logging_utils.add_logging_options(parser)
     args = parser.parse_args()
@@ -91,15 +92,6 @@ def main():
     msg = "Reading annotated transcripts"
     logger.info(msg)
     annotated_transcripts = bed_utils.read_bed(args.annotated_transcripts)
-
-    msg = "Reading extracted ORFs and exons"
-    logger.info(msg)
-    extracted_orfs = bed_utils.read_bed(args.extracted_orfs)
-    extracted_orf_exons = bed_utils.read_bed(args.orf_exons)
-
-    msg = "Found {} extracted ORFs with {} exons".format(len(extracted_orfs),
-                                                         len(extracted_orf_exons))
-    logger.debug(msg)
 
     # get the annotated transcript exons
     if args.annotated_exons is None:
@@ -115,12 +107,28 @@ def main():
 
         annotated_exons = bed_utils.read_bed(args.annotated_exons)
 
+    msg = "Reading extracted ORFs"
+    logger.info(msg)
+    extracted_orfs = bed_utils.read_bed(args.extracted_orfs)
+
+    if args.orf_exons is None:
+        msg = "Splitting the extracted ORFs into exon blocks"
+        logger.info(msg)
+        extracted_orf_exons = bed_utils.split_bed12(extracted_orfs,
+                                                    num_cpus=args.num_cpus,
+                                                    progress_bar=True)
+    else:
+        msg = "Reading the extracted ORFs exons"
+        logger.info(msg)
+        extracted_orf_exons = bed_utils.read_bed(args.orf_exons)
+
+    msg = "Found {} extracted ORFs with {} exons".format(len(extracted_orfs),
+                                                         len(extracted_orf_exons))
+    logger.debug(msg)
+
     # filter out the ORFs that are entirely within annotated transcripts
     if args.filter:
-        msg = ("Removing extracted ORFs which are completely covered by the "
-               "annotated transcripts")
-        logger.info(msg)
-        msg = "Finding completely covered extracted ORFs"
+        msg = "Removing ORFs which are completely covered by the annotated transcripts"
         logger.info(msg)
 
         nonoverlapping_ids = bed_utils.subtract_bed(extracted_orf_exons,
@@ -135,7 +143,7 @@ def main():
         msg = "After filtering, {} extracted ORFs remain".format(len(extracted_orfs))
         logger.info(msg)
 
-    # annotate and remove the ORFs which do not at all overlap with the annotations
+    # annotate and remove the ORFs which do not at all overlap the annotations
     if args.nonoverlapping_label is not None:
         nonoverlapping_ids = bed_utils.subtract_bed(extracted_orfs,
                                                     annotated_transcripts,
@@ -202,7 +210,7 @@ def main():
                                             num_cpus=args.num_cpus,
                                             progress_bar=True)
 
-    # First, remove all in-frame (canonical, canonical variants), and also within (oof) ORFs
+    # First, remove all in-frame (canonical, canonical variants), and also within and oof ORFs
 
     msg = "Marking canonical and extracted ORFs with the same stop codon"
     logger.info(msg)
@@ -235,11 +243,17 @@ def main():
                                                min_a_overlap=1,
                                                min_b_overlap=1)
 
-    exact_match_orf_ids = {m.b_info for m in exact_matches}
+    exact_match_pairs = {(m.a_info, m.b_info) for m in exact_matches}
+    exact_match_orf_ids = {pair[1] for pair in exact_match_pairs}
+
     m_exact_orf_matches = extracted_orf_exons['id'].isin(exact_match_orf_ids)
     extracted_orf_exons = extracted_orf_exons[~m_exact_orf_matches]
+
     m_canonical = extracted_orfs['id'].isin(exact_match_orf_ids)
-    extracted_orfs.loc[m_canonical, 'orf_type'] = 'canonical'
+    label = 'canonical'
+    extracted_orfs.loc[m_canonical, 'orf_type'] = label
+
+    type_to_matches = {label: exact_match_pairs}
 
     msg = "Found {} canonical ORFs".format(len(exact_match_orf_ids))
     logger.info(msg)
@@ -252,18 +266,18 @@ def main():
                                                   min_a_overlap=1)
 
     # make sure the "end"s match before calling something an extended match
-    extended_match_pairs = {(m.a_info, m.b_info) for m in tqdm.tqdm(extended_matches)
+    extended_match_pairs = {(m.a_info, m.b_info) for m in extended_matches
                             if (m.a_info, m.b_info) in canonical_extracted_matching_ends}
     extended_match_ids = {pair[1] for pair in extended_match_pairs}
 
     m_extended_matches = extracted_orf_exons['id'].isin(extended_match_ids)
     extracted_orf_exons = extracted_orf_exons[~m_extended_matches]
-    m_canonical_extended = extracted_orfs['id'].isin(extended_match_ids)
 
+    m_canonical_extended = extracted_orfs['id'].isin(extended_match_ids)
     label = "{}canonical_extended".format(args.label_prefix)
     extracted_orfs.loc[m_canonical_extended, 'orf_type'] = label
 
-    type_to_matches = {label: extended_match_pairs}
+    type_to_matches[label] = extended_match_pairs
 
     msg = "Found {} canonical_extended ORFs".format(len(extended_match_ids))
     logger.info(msg)
@@ -276,14 +290,14 @@ def main():
                                                    min_b_overlap=1)
 
     # make sure the "end"s match before calling something a truncated match
-    truncated_match_pairs = {(m.a_info, m.b_info) for m in tqdm.tqdm(truncated_matches)
+    truncated_match_pairs = {(m.a_info, m.b_info) for m in truncated_matches
                              if (m.a_info, m.b_info) in canonical_extracted_matching_ends}
     truncated_match_ids = {pair[1] for pair in truncated_match_pairs}
 
     m_truncated_matches = extracted_orf_exons['id'].isin(truncated_match_ids)
     extracted_orf_exons = extracted_orf_exons[~m_truncated_matches]
-    m_canonical_truncated = extracted_orfs['id'].isin(truncated_match_ids)
 
+    m_canonical_truncated = extracted_orfs['id'].isin(truncated_match_ids)
     label = "{}canonical_truncated".format(args.label_prefix)
     extracted_orfs.loc[m_canonical_truncated, 'orf_type'] = label
 
@@ -292,13 +306,11 @@ def main():
     msg = "Found {} canonical_truncated ORFs".format(len(truncated_match_ids))
     logger.info(msg)
 
-    msg = ("Labeling ORFs which are completely covered by a canonical ORF but "
-           "are out-of-frame")
+    msg = ("Finding ORFs which are within canonical ORFs but "
+           "do not share the annotated stop codon (e.g. in frame stop, "
+           "out-of-frame)")
     logger.info(msg)
 
-    # anything in "truncated matches" which *does not* share a stop codon with
-    # the match is a "within" ORF ... not necessarily, see
-    # https://github.com/dieterich-lab/rp-bp/issues/96
     within_pairs = {(m.a_info, m.b_info) for m in truncated_matches
                     if m.b_info not in truncated_match_ids}
     within_ids = {pair[1] for pair in within_pairs}
@@ -313,14 +325,14 @@ def main():
 
         msg = "Found {} within ORFs".format(len(within_ids))
         logger.info(msg)
-    else:
+    else:  # but there should not be any novel within...
         label = "{}overlap".format(args.label_prefix)
 
     extracted_orfs.loc[m_within, 'orf_type'] = label
 
     # then, find ORFs that are overlapping, or that are entirely within the UTRs
 
-    msg = "Finding all overlap matches"
+    msg = "Finding all UTR overlap matches"
     logger.info(msg)
     out_of_frame_matches = bed_utils.get_bed_overlaps(canonical_orf_exons,
                                                       extracted_orf_exons)
@@ -335,20 +347,19 @@ def main():
            "canonical ORF and annotated leaders or trailers")
     logger.info(msg)
 
-    # make sure up-/downstream overlaps are unique,
-    # in cases where an ORF overlaps both the 5'UTR and
-    # the 3'UTR of different same sense overlapping transcripts,
-    # it is assigned by default to the downstream overlap
-    # for de novo, everything is labeled as overlap
+    # We need to choose how to ensure that up-/downstream overlaps are unique.
+    # Where an ORF overlaps both the 5'UTR and the 3'UTR of different same
+    # sense overlapping transcripts, it is assigned by default to the downstream overlap.
+    # For de novo, everything is labeled as overlap.
 
     leader_match_pairs = {(m.a_info, m.b_info) for m in leader_matches}
     trailer_match_pairs = {(m.a_info, m.b_info) for m in trailer_matches}
 
     if args.nonoverlapping_label is None:
 
-        # for standard assembly, we also need to make sure
+        # For standard assembly, we also need to make sure that
         # all overlap matches are fully contained within the
-        # transcript structure
+        # transcript structure.
         transcript_matches = bed_utils.get_bed_overlaps(annotated_exons,
                                                         extracted_orf_exons,
                                                         min_b_overlap=1)
@@ -365,9 +376,9 @@ def main():
                                  and (m.a_info, m.b_info) not in leader_match_pairs
                                  and (m.a_info, m.b_info) in transcript_match_pairs}
 
-        # we do not assign preference in cases where the ORF overlaps both sides
+        # We do not assign preference where the ORF overlaps both sides
         # of the coding sequence on the same transcript, any ORF
-        # satisfying both will be labeled simply as overlap
+        # satisfying both will be labeled simply as overlap.
         overlap_pairs = {(m.a_info, m.b_info) for m in out_of_frame_matches
                          if (m.a_info, m.b_info) in leader_match_pairs
                          and (m.a_info, m.b_info) in trailer_match_pairs
@@ -390,14 +401,14 @@ def main():
         m_leader_overlap_matches = extracted_orf_exons['id'].isin(leader_overlap_ids)
         extracted_orf_exons = extracted_orf_exons[~m_leader_overlap_matches]
 
-        m_trailer_overlap_matches = extracted_orf_exons['id'].isin(trailer_overlap_ids)
-        extracted_orf_exons = extracted_orf_exons[~m_trailer_overlap_matches]
-
         m_five_prime_overlap = extracted_orfs['id'].isin(leader_overlap_ids)
         label = "{}five_prime_overlap".format(args.label_prefix)
         extracted_orfs.loc[m_five_prime_overlap, 'orf_type'] = label
 
         type_to_matches[label] = leader_overlap_pairs
+
+        m_trailer_overlap_matches = extracted_orf_exons['id'].isin(trailer_overlap_ids)
+        extracted_orf_exons = extracted_orf_exons[~m_trailer_overlap_matches]
 
         m_three_prime_overlap = extracted_orfs['id'].isin(trailer_overlap_ids)
         label = "{}three_prime_overlap".format(args.label_prefix)
@@ -433,9 +444,8 @@ def main():
     msg = "Finding ORFs completely within 5' or 3' leaders"
     logger.info(msg)
 
-    # here, upstream ORFs are labeled first and immediately removed,
-    # so there is no double match, we just need to make sure below
-    # that the transcript part of the "orf_id" is consistent with the label.
+    # Upstream ORFs are labeled first and immediately removed,
+    # so there is no double match, we just need to make sure below.
 
     leader_matches = bed_utils.get_bed_overlaps(five_prime_exons,
                                                 extracted_orf_exons,
@@ -511,28 +521,20 @@ def main():
     msg = "Found {} unlabeled ORFs".format(sum(m_no_orf_type))
     logger.info(msg)
 
-    msg = "Now checking all labels against orf_ids"
+    msg = "Now assign transcript id to ORFs"
     logger.info(msg)
 
-    if not args.skip_check:
+    grouped = extracted_orfs.groupby('orf_type')
+    extracted_orfs = parallel.apply_parallel_groups(grouped, args.num_cpus,
+                                                    assign_transcripts, type_to_matches,
+                                                    progress_bar=True)
+    extracted_orfs = pd.concat(extracted_orfs)
 
-        grouped = extracted_orfs.groupby('orf_type')
-        grouped_canonical = grouped.get_group('canonical')
-        orfs_group = extracted_orfs.drop(grouped_canonical.index).groupby('orf_type')
-
-        relabeled_orfs = parallel.apply_parallel_groups(orfs_group, args.num_cpus,
-                                                        check_orf_ids, type_to_matches,
-                                                        progress_bar=True)
-        relabeled_orfs.append(grouped_canonical)
-        extracted_orfs = pd.concat(relabeled_orfs)
-
-    msg = "Writing ORFs with types to disk"
+    msg = "Writing ORFs with labels to disk"
     logger.info(msg)
 
     extracted_orfs = bed_utils.sort(extracted_orfs)
     additional_columns = ['orf_num', 'orf_len', 'orf_type']
-    if 'assoc_trx' in extracted_orfs.columns:
-        additional_columns.extend(['assoc_trx'])
     fields = bed_utils.bed12_field_names + additional_columns
     extracted_orfs = extracted_orfs[fields]