+++ rpbp/reference_preprocessing/prepare_rpbp_genome.py
@@ -1,5 +1,11 @@
 #! /usr/bin/env python3
 
+"""This is the main script used to create the reference
+genome indices, identify and label the ORFs.
+
+Calls: extract-orf-coordinates and label-orfs
+"""
+
 import os
 import sys
 import yaml
@@ -21,17 +27,15 @@ logger = logging.getLogger(__name__)
 
 
 def get_orfs(gtf, args, config, is_annotated=False, is_de_novo=False):
-    """ This helper function processes a GTF file into its ORFs.
+    """ Process a GTF file into its ORFs.
     """
+
     call = not args.do_not_call
     chr_name_file = os.path.join(config['star_index'], 'chrName.txt')
     chr_name_str = "--chr-name-file {}".format(chr_name_file)
 
     logging_str = logging_utils.get_logging_options_string(args)
     cpus_str = "--num-cpus {}".format(args.num_cpus)
-    add_option_str = ''
-    if args.add_trx_match:
-        add_option_str = '--add-trx-match'
 
     # extract a BED12 of the annotated ORFs
     transcript_bed = filenames.get_bed(config['genome_base_path'],
@@ -50,20 +54,6 @@ def get_orfs(gtf, args, config, is_annotated=False, is_de_novo=False):
     shell_utils.call_if_not_exists(cmd, out_files, in_files=in_files,
                                    overwrite=args.overwrite, call=call)
 
-    exons_file = filenames.get_exons(config['genome_base_path'],
-                                     config['genome_name'],
-                                     is_annotated=is_annotated,
-                                     is_de_novo=is_de_novo)
-
-    cmd = ("split-bed12-blocks {} {} --num-cpus {} {}".format(transcript_bed,
-                                                              exons_file,
-                                                              args.num_cpus,
-                                                              logging_str))
-    in_files = [transcript_bed]
-    out_files = [exons_file]
-    shell_utils.call_if_not_exists(cmd, out_files, in_files=in_files,
-                                   overwrite=args.overwrite, call=call)
-
     # extract the transcript fasta
     transcript_fasta = filenames.get_transcript_fasta(config['genome_base_path'],
                                                       config['genome_name'],
@@ -79,7 +69,7 @@ def get_orfs(gtf, args, config, is_annotated=False, is_de_novo=False):
     shell_utils.call_if_not_exists(cmd, out_files, in_files=in_files,
                                    overwrite=args.overwrite, call=call)
 
-    # extract ORFs using genomic coordinates
+    # extract ORFs from the transcripts using genomic coordinates
     orfs_genomic = filenames.get_orfs(config['genome_base_path'],
                                       config['genome_name'],
                                       note=config.get('orf_note'),
@@ -94,29 +84,13 @@ def get_orfs(gtf, args, config, is_annotated=False, is_de_novo=False):
                                                                 cpus_str,
                                                                 start_codons_str,
                                                                 stop_codons_str,
-                                                                logging_str,
-                                                                add_option_str)
+                                                                logging_str)
     in_files = [transcript_fasta, transcript_bed]
     out_files = [orfs_genomic]
     shell_utils.call_if_not_exists(cmd, out_files, in_files=in_files,
                                    overwrite=args.overwrite, call=call)
 
-    exons_file = filenames.get_exons(config['genome_base_path'],
-                                     config['genome_name'],
-                                     note=config.get('orf_note'),
-                                     is_annotated=is_annotated,
-                                     is_de_novo=is_de_novo,
-                                     is_orf=True)
-
-    cmd = ("split-bed12-blocks {} {} --num-cpus {} {}".format(orfs_genomic,
-                                                              exons_file,
-                                                              args.num_cpus,
-                                                              logging_str))
-    in_files = [orfs_genomic]
-    out_files = [exons_file]
-    shell_utils.call_if_not_exists(cmd, out_files, in_files=in_files,
-                                   overwrite=args.overwrite, call=call)
-
+    # label the ORFs
     labeled_orfs = orfs_genomic
     annotated_bed = filenames.get_bed(config['genome_base_path'],
                                       config['genome_name'],
@@ -124,52 +98,48 @@ def get_orfs(gtf, args, config, is_annotated=False, is_de_novo=False):
                                       is_annotated=True)
 
     de_novo_str = ""
-    overwrite_exons = True
     if is_de_novo:
-        de_novo_str = '--label-prefix "novel_" --filter --nonoverlapping-label "novel" -s'
-        overwrite_exons = False
+        de_novo_str = '--label-prefix "novel_" --filter --nonoverlapping-label "novel"'
 
-    cmd = "label-orfs {} {} {} {} {} {} {}".format(annotated_bed,
-                                                   orfs_genomic,
-                                                   exons_file,
-                                                   labeled_orfs,
-                                                   cpus_str,
-                                                   de_novo_str,
-                                                   logging_str)
-    in_files = [annotated_bed, orfs_genomic, exons_file]
-    #  since we are reusing the name, it will already exist
-    out_files = None  # [] # [labeled_orfs]
+    cmd = "label-orfs {} {} {} {} {} {}".format(annotated_bed,
+                                                orfs_genomic,
+                                                labeled_orfs,
+                                                de_novo_str,
+                                                logging_str,
+                                                cpus_str)
+    in_files = [annotated_bed, orfs_genomic]
+    #  overwrites the input file
+    out_files = None
     shell_utils.call_if_not_exists(cmd, out_files, in_files=in_files,
                                    overwrite=args.overwrite, call=call)
 
-    # after checking orf ids, we need to re-write the exons, unless [--skip-check]
-    # has been used, with de novo
-    cmd = ("split-bed12-blocks {} {} --num-cpus {} {}".format(orfs_genomic,
+    # now write the ORF exons
+    exons_file = filenames.get_exons(config['genome_base_path'],
+                                     config['genome_name'],
+                                     note=config.get('orf_note'),
+                                     is_annotated=is_annotated,
+                                     is_de_novo=is_de_novo)
+
+    cmd = ("split-bed12-blocks {} {} --num-cpus {} {}".format(labeled_orfs,
                                                               exons_file,
                                                               args.num_cpus,
                                                               logging_str))
-    in_files = [orfs_genomic]
+    in_files = [labeled_orfs]
     out_files = [exons_file]
     shell_utils.call_if_not_exists(cmd, out_files, in_files=in_files,
-                                   overwrite=overwrite_exons, call=call)
+                                   overwrite=args.overwrite, call=call)
 
 
 def main():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-                                     description='''This script creates all of the files 
-                                     necessary for downstream analysis performed with 
-                                     the rpbp package.''')
+                                     description='''Prepare a reference genome and matching 
+        annotations, including labelled ORFs, for use with the Rp-Bp periodicity estimation 
+        and ORF translation prediction pipeline.''')
 
-    parser.add_argument('config', help='''The (yaml) config file''')
+    parser.add_argument('config', help='''The (yaml) configuration file''')
 
     parser.add_argument('--overwrite', help='''If this flag is present, existing files
         will be overwritten.''', action='store_true')
-
-    parser.add_argument('--add-trx-match', help='''If this flag is present, an additional
-        column is added to the ORFs file containing for each ORF a list of annotated 
-        transcripts to which it belongs. This is not used as part of the Rp-Bp pipeline, 
-        but may be useful for downstream analysis, this however significantly increase 
-        the running time to extract the ORFs.''', action='store_true')
 
     star_utils.add_star_options(parser)
     slurm.add_sbatch_options(parser)
@@ -239,18 +209,24 @@ def main():
 
     # eventually, we will use these names
     annotated_orfs = filenames.get_orfs(config['genome_base_path'],
-                                        config['genome_name'], note=config.get('orf_note'), is_annotated=True,
+                                        config['genome_name'],
+                                        note=config.get('orf_note'),
+                                        is_annotated=True,
                                         is_de_novo=False)
 
     annotated_exons_file = filenames.get_exons(config['genome_base_path'],
-                                               config['genome_name'], note=config.get('orf_note'),
-                                               is_annotated=True, is_de_novo=False, is_orf=True)
+                                               config['genome_name'],
+                                               note=config.get('orf_note'),
+                                               is_annotated=True,
+                                               is_de_novo=False)
 
     orfs_genomic = filenames.get_orfs(config['genome_base_path'],
-                                      config['genome_name'], note=config.get('orf_note'))
+                                      config['genome_name'],
+                                      note=config.get('orf_note'))
 
     exons_file = filenames.get_exons(config['genome_base_path'],
-                                     config['genome_name'], note=config.get('orf_note'), is_orf=True)
+                                     config['genome_name'],
+                                     note=config.get('orf_note'))
 
     use_gff3_specs = config['gtf'].endswith('gff')
     gtf_file = filenames.get_gtf(config['genome_base_path'],
@@ -258,17 +234,20 @@ def main():
 
     # now, check if we have a de novo assembly
     if 'de_novo_gtf' in config:
-        get_orfs(config['de_novo_gtf'], args, config, is_annotated=False,
-                 is_de_novo=True)
+        get_orfs(config['de_novo_gtf'], args, config, is_annotated=False, is_de_novo=True)
 
         # we need to concat the ORF and exon files
         de_novo_orfs = filenames.get_orfs(config['genome_base_path'],
-                                          config['genome_name'], note=config.get('orf_note'), is_annotated=False,
+                                          config['genome_name'],
+                                          note=config.get('orf_note'),
+                                          is_annotated=False,
                                           is_de_novo=True)
 
         de_novo_exons_file = filenames.get_exons(config['genome_base_path'],
-                                                 config['genome_name'], note=config.get('orf_note'),
-                                                 is_annotated=False, is_de_novo=True, is_orf=True)
+                                                 config['genome_name'],
+                                                 note=config.get('orf_note'),
+                                                 is_annotated=False,
+                                                 is_de_novo=True)
 
         orfs_files = [annotated_orfs, de_novo_orfs]
 
@@ -281,8 +260,6 @@ def main():
             concatenated_bed = bed_utils.concatenate(orfs_files, sort_bed=True)
             concatenated_bed['orf_num'] = range(len(concatenated_bed))
             additional_columns = ['orf_num', 'orf_len', 'orf_type']
-            if 'assoc_trx' in concatenated_bed.columns:
-                additional_columns.extend(['assoc_trx'])
             fields = bed_utils.bed12_field_names + additional_columns
             bed_utils.write_bed(concatenated_bed[fields], orfs_genomic)
         else:
@@ -321,7 +298,7 @@ def main():
                 shell_utils.create_symlink(config['gtf'], gtf_file, call)
 
     else:
-        # finally, make sure our files are named correctly
+        # if we do not have a de novo assembly, symlink the files
 
         if os.path.exists(annotated_orfs):
             shell_utils.create_symlink(annotated_orfs, orfs_genomic, call)